Move macro definitions and helpers to their own ubmodule
